Add conftest file and move fixtures
